Update blackjack.rb
Good job leveraging classes for this assignment. Be careful with your capitalization and watch to make sure method and class is "closed" (with and end) at the spot you want it to. I made a few comments of areas that could be improved and to point out errors you may be having. I hope this helps. I look forward to trying out your game once it is up and running. 

(Don't worry about making these changes this weekend. We will talk more about this on Monday)